pyproject.toml
[tool.poetry]
name = "site-api"
version = "0.1.0"
description = ""
authors = ["Mikhail Seliakov <selyakov.md@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.0"}
pytest = "^7.4.4"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
pydantic = {extras = ["email"], version = "^2.5.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
setuptools = "^69.0.3"
gunicorn = "^21.2.0"
requests = "^2.31.0"
fastapi-users = "^12.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
